backend/requirements.txt
Django==4.1.5
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
Pillow==9.4.0
psycopg2-binary==2.9.5
sqlparse==0.4.3
asgiref==3.6.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4